+++ src/main/java/hu/ibello/apitest/HttpClient.java
@@ -98,8 +98,9 @@ public interface HttpClient {
 	 * Sets the body of the HTTP request.
 	 * The parameter can be:
 	 * <ul>
-	 * <li>a {@link File} instance, in this case a multipart attachment will be added to the body,</li>
+	 * <li>a {@link File} instance, in this case a multipart/form-data attachment will be added to the body,</li>
 	 * <li>an {@link InputStream} instance, in this case the content of the stream will be the request body,</li>
+	 * <li>a byte array, in this case the bytes will be the request body,</li>
 	 * <li>or anything else - which will be transformed to a {@link String}.</li>
 	 * </ul>
 	 * @param object the object to be sent
@@ -107,6 +108,14 @@ public interface HttpClient {
 	 * @return this {@link HttpClient} instance
 	 */
 	public <T> HttpClient body(T object);
+	
+	/**
+	 * Sets the name of the body.
+	 * If specified, the request will be multipart/form-data type, and the body will be added with this name.
+	 * @param name the name of the body
+	 * @return this {@link HttpClient} instance
+	 */
+	public HttpClient name(String name);
 	
 	/**
 	 * Adds a HTTP header to the request.