+++ src/main/markdown/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ibello-api changelog
 
+## 1.24.0
+
+- Minimum required java version: 11
+
 ## 1.23.0
 
 2024-10-15